Refactored discourse app facade.py
part of #1680
